+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- education: add EducationClient + courses/enrollments services and sdk.education getter for the education domain. @fdaei
+
 ## [1.3.15] - 2026-06-25
 
 ### Added